requirements.txt
# requirements.txt

flask
flask-socketio
google-genai
python-dotenv